new move of modal and eventlistener
